refactor(viewmodel): separate MovieListViewModel and MovieDetailsViewModel
- Created MovieListViewModel to manage movie list state and API calls.
- Created MovieDetailsViewModel to manage movie details state and API calls.
- Refactored MovieListScreen and MovieDetailsScreen to use their respective ViewModels.
