+++ app/src/main/java/com/daniellapaiva/movieapp/di/AppModule.kt
@@ -4,7 +4,8 @@ import com.daniellapaiva.movieapp.data.repository.MovieRepositoryImpl
 import com.daniellapaiva.movieapp.domain.usecase.GetPopularMoviesUseCase
 import com.daniellapaiva.movieapp.domain.MovieRepository
 import com.daniellapaiva.movieapp.domain.usecase.GetMovieDetailsUseCase
-import com.daniellapaiva.movieapp.presentation.MovieViewModel
+import com.daniellapaiva.movieapp.presentation.viewmodel.MovieDetailViewModel
+import com.daniellapaiva.movieapp.presentation.viewmodel.MovieListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -13,8 +14,8 @@ val appModule = module {
     single<MovieRepository> { MovieRepositoryImpl(get()) }
 
     single { GetPopularMoviesUseCase(get()) }
-
     single { GetMovieDetailsUseCase(get()) }
 
-    viewModel { MovieViewModel(get(), get()) }
+    viewModel { MovieListViewModel(get()) }
+    viewModel { MovieDetailViewModel(get()) }
 }
