+++ app/src/main/java/com/daniellapaiva/movieapp/presentation/viewmodel/MovieListViewModel.kt
@@ -1,0 +1,28 @@
+package com.daniellapaiva.movieapp.presentation.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.daniellapaiva.movieapp.domain.model.Movie
+import com.daniellapaiva.movieapp.domain.usecase.GetPopularMoviesUseCase
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.launch
+
+class MovieListViewModel(
+    private val getPopularMoviesUseCase: GetPopularMoviesUseCase,
+) : ViewModel() {
+
+    private val _popularMovies = MutableStateFlow<List<Movie>>(emptyList())
+    val popularMovies: StateFlow<List<Movie>> = _popularMovies
+
+    init {
+        fetchPopularMovies()
+    }
+
+    private fun fetchPopularMovies() {
+        viewModelScope.launch {
+            val movies = getPopularMoviesUseCase()
+            _popularMovies.value = movies
+        }
+    }
+}
